refactor(foundation-demo): add additional directory level to separate checkout/components payment and shipping
add additional directory level to separate checkout/components payment and shipping
