+++ apps/foundation-demo/src/app/checkout/checkout.module.ts
@@ -1,17 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CheckoutViewComponent } from './pages/checkout-view/checkout-view.component';
-import { ShippingFormComponent } from './components/shipping-form/shipping-form.component';
+import { ShippingFormComponent } from './components/shipping/shipping-form/shipping-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { CoreCheckoutModule } from '@daffodil/core';
-import { ShippingSummaryComponent } from './components/shipping-summary/shipping-summary.component';
-import { ShippingOptionsComponent } from './components/shipping-options/shipping-options.component';
-import { ShippingComponent } from './components/shipping/shipping.component';
+import { ShippingSummaryComponent } from './components/shipping/shipping-summary/shipping-summary.component';
+import { ShippingOptionsComponent } from './components/shipping/shipping-options/shipping-options.component';
+import { ShippingComponent } from './components/shipping/shipping/shipping.component';
 import { DesignModule } from '../design/design.module';
 import { FoundationCheckoutStateModule } from './checkout-state.module';
-import { PaymentFormComponent } from './components/payment-form/payment-form.component';
-import { PaymentComponent } from './components/payment/payment.component';
-import { PaymentSummaryComponent } from './components/payment-summary/payment-summary.component';
+import { PaymentFormComponent } from './components/payment/payment-form/payment-form.component';
+import { PaymentComponent } from './components/payment/payment/payment.component';
+import { PaymentSummaryComponent } from './components/payment/payment-summary/payment-summary.component';
 import { CartModule } from '../cart/cart.module';
 
 @NgModule({
