+++ apps/foundation-demo/src/app/checkout/components/payment/payment-form/payment-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Validators, FormGroup, FormBuilder, FormControl } from '@angular/forms';
 import { PaymentInfo } from '@daffodil/core';
-import { ErrorStateMatcher } from '../../../design/molecules/error-state-matcher/error-state-matcher.component';
+import { ErrorStateMatcher } from '../../../../design/molecules/error-state-matcher/error-state-matcher.component';
 
 @Component({
   selector: 'payment-form',
