+++ apps/foundation-demo/src/app/checkout/components/payment/payment-summary/payment-summary.component.scss
