+++ apps/foundation-demo/src/app/checkout/components/payment/payment/payment.component.ts
@@ -1,14 +1,13 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { PaymentInfo } from '@daffodil/core';
 import { Store, select } from '@ngrx/store';
-import * as fromFoundationCheckout from '../../reducers';
-import { SetShowPaymentForm, ToggleShowPaymentForm } from '../../actions/payment.actions';
+import * as fromFoundationCheckout from '../../../reducers';
+import { SetShowPaymentForm, ToggleShowPaymentForm } from '../../../actions/payment.actions';
 import { Observable } from 'rxjs';
 
 @Component({
   selector: 'payment',
-  templateUrl: './payment.component.html',
-  styleUrls: ['./payment.component.scss']
+  templateUrl: './payment.component.html'
 })
 export class PaymentComponent implements OnInit {
