--- apps/foundation-demo/src/app/checkout/components/shipping-async-wrapper/shipping-async-wrapper.component.html
@@ -1,3 +1,0 @@
-<div class="checkout__shipping">
-  <shipping [isShippingInfoValid]="isShippingInfoValid" (continueToPayment)="onContinueToPayment()"></shipping>
-</div>
