--- apps/foundation-demo/src/app/checkout/components/shipping-async-wrapper/shipping-async-wrapper.component.ts
@@ -1,15 +1,0 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
-
-@Component({
-  selector: 'shipping-async-wrapper',
-  templateUrl: './shipping-async-wrapper.component.html'
-})
-export class ShippingAsyncWrapperComponent {
-
-  @Input() isShippingInfoValid: Boolean;
-  @Output() continueToPayment: EventEmitter<any> = new EventEmitter();
-
-  onContinueToPayment() {
-    this.continueToPayment.emit();
-  }
-}
